[COMMON] cpuidle: skip correcting predicted idle time
Menu goernor corrects predicted idle time on the tendency of past
residency time. It is for cpuidle optimization, there are
exception. For example, if the cpus in a cluster are working busily
and they become idle suddenly, we expect that cluster is powered
down but just cpus are down by correcting idle time. To avoid
this situation, add skip_correction field.

In exynos, skip_correction is enabled to non-boot cluster.

Change-Id: Ice316e5bb791ba1c67cc343268b8641de95c89a0
Signed-off-by: Park Bumgyu <bumgyu.park@samsung.com>