[COMMON] PCI: exynos: Add support for history buffer
When the power state is changed, the information of power states
such as LTSSM (Link training and status state machine) state,
L1 sub state and D state is stored to history buffers. This buffer
is enabled when link is up, and the history buffers are printed
when link down interrupt happened or link up failed. The information
of power states is helpful for debugging link down issues.

Change-Id: I177633ac2a99b6bba588a108470d8ed22e4701ba
Suggested-by: Kyoungil Kim <ki0351.kim@samsung.com>
Signed-off-by: Jingoo Han <jg1.han@samsung.com>