[COMMON] media: mfc: Check for hw_lock of all the contexts before sleep
The current logic checks only the current ctx's hw_lock before calling
the sleep function and inturn issuing SLEEP cmd to firmware.

But in a multi instance scenario (decoding + encoding), it can so happen that
another ctx may also be running and we issue SLEEP cmd to firmware without
checking all the ctx's hw_lock.
In such cases, the behaviour of the firmware is not gauranteed and
can throw errors.
Checking for hw_lock of all the contexts before issuing sleep cmd
avoids the above mentioned case.

Change-Id: I014ba8abb86173af94ee32a729f14fa13e249f69
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Signed-off-by: Vikas Sajjan <vikas.sajjan@samsung.com>