[COMMON] media: vb2-ion: fix insufficient scope check
vb2_ion_sync_for_cpu() and vb2_ion_sync_for_device() provides cache
maintenance for the client drivers especially for the drivers that
allocates from vb2-ion. If a device driver generated a kernel address
from an allocated buffer, it can ask vb2-ion for maintaining caches
about a part of the buffer with offset and length which implies the
area of interest.
The length and the offset are calculated in a client driver which is
run in the kernelspace. Therefore, they shoud be correct.

This patch which generates BUG on invalid scope information does not
have mercy on the client drivers because invalid data handled in the
drivers should be fixed.

Change-Id: Idd9e8cec5acff48562838e29a2f7c0a301d6d5bf
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>