[common] fbdev: vpp: Fix vpp_set_read_order for supporting bot evt0 and evt1
VGR0/VGR1 have a rotation function. When the H/Ws are running the function,
the memory access direction also can be changed. If that, not only master IPs
but also the direction of sysmmu prefetch buffer(pb) should be changed.
The sysmmu driver supports a api which is sysmmu_set_prefetch_buffer_property.
Unfortunately, the configurations of VGR pb in evt0 are different from its in evt1.
So, device tree file and vpp driver should be changed for supporting evt0 and evt1.

Change-Id: I2cb25ab008bef2d2fd4970f58273c562e609c991
Signed-off-by: Sungchun Kang <sungchun.kang@samsung.com>