usb: s3c-hsotg: Powering down USB LDOs, Phy and clock in disconnect state
Vbus detect gpio available on boards can be used to detect any
activity on that line (a USB cable connect/disconnect). We can
use the same to power down the regulators, usb-phy and usb clocks.
So, turning these off when usb gadget cable is not connected and
again turning them back when gadget cable is connected and a
high VBus is detected on the line.

Also added a config symbol CONFIG_EXYNOS3_USB_REGULATOR_DISABLE to
control this regulator turning-off.
The USB regulators (ldo-6 and ldo-8) are also used by other
IPs. LDO-6 is used by MIPI and is always turned on. But LDO-8
which is also the regulator for main UART on koi can be turned off
when USB is not connected and we don't need any serial for debugging
purposes.
Disable this config, if we want UART to be available even after USB
is disconnected.

Change-Id: I8c653cb1487088d933366aec3c405e0dad7ebaea
Signed-off-by: Chandni <ch.chandni@samsung.com>
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>