[COMMON] media: m2m1shot2: check if cache maintenance is really required
Userspace can inform the driver that the cache maintenance is
unneccessary if it knows how CPU accesses to the buffer. If cache
maintenance is not required before DMA, the userspace gives
M2M1SHOT2_IMGFLAG_NO_CACHECLEAN while it gives
M2M1SHOT2_IMGFLAG_NO_CACHEINV if cache invalidation is not required
after DMA.

For the userptr, it is possible to determine if the buffer is mapped
in the userspace with cached attribute. If it does not have cached
attribute, cache maintenance is not required.

Change-Id: I186f25299c899ba70b7504e89115db0421e1762a
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>