[COMMON] media: m2m1shot2: call .device_run() in a process context
If .device_run() of a context called after waiting an acquire fence,
it is called in IRQ disabled context because the callback called
by the fence waiter is called in IRQ disabled context. But .device_run
should be called in a process context unless it is called by the
IRQ handler of the client device driver because .device_run() may call
blocking function including pm_runtime_get_sync().

Change-Id: If54aa6e85972c83fb5957f6a0b040eff6092be40
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>