[7870] PM: DEVFREQ: change frequency scale function
Previously, the devfreq drivers were using the CAL APIs
for faster processing when frequency scale. However,
if the gating method is using in the bus or top domains,
the synchronization problems can be occur becasue the dfs CAL APIs
are not using synchronize function. So, the scale function
is changed to CCF.

Change-Id: I0726ba29a73b732e7ab6b3c8c88a272b3c78c27c
Signed-off-by: Sungjinn Chung <sungjinn.chung@samsung.com>