[COMMON] cpuidle: disable cpuidle during hotplug
During hotplug, entry cpuidle and cluster power down increases latency.
If core to be offed is in the idle, needs to wake-up time for task what have to
be executed on the core to be offed and it increases latency.
To reduce latency, cpuidle is disabled between CPU_DOWN_PREPARE and CPU_ONLINE notify.

Change-Id: I7eb88cbd678fb860caa2e271ec69b1da4755e033
Signed-off-by: Youngtae Lee <yt0729.lee@samsung.com>