[8890] video: exynos: decon: screen clear with window colormap
Garbage is displayed on S6E3HF4 LCD because of DSC compression.
The DECON transfers black frame to LCD when LCD is turned on or off.
DSC decoder is enabled in DDI but DECON sends uncompressed data.
It will make garbage display.

So, compressed black frame using winmap is transferred to LCD.

Change-Id: Ic62a2ee11b2adaff0daad157392995b59c5bddf0
Signed-off-by: SeWoon Park <seuni.park@samsung.com>