arm: koi_defconfig: Enable CONFIG_EXYNOS3_USB_REGULATOR_DISABLE
CONFIG_EXYNOS3_USB_REGULATOR_DISABLE will let the USB reguators to be
turned-off control this regulator turning-off when USB cable is not
connected.
The USB regulators (ldo-6 and ldo-8) are also used by other
IPs. LDO-6 is used by MIPI and is always turned on. But LDO-8,
which is also the regulator for main UART on koi, can be turned off
when USB cable is disconnected and we don't want any serial output.

Disable this config, if we want UART to be available even after USB
is disconnected.

Change-Id: If238b8b607f1df1b4b492dc954bfad790db04292
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>