[COMMON] media: m2m1shot2: wait for pending tasks on context release
The release of a context may be called for a pending tasks that is
waiting to be serviced by the H/W. For the cases, m2m1shot2 now waits
until the pending task is also finished when the context is being
destroyed.

This change is made for the users that want to forget about releasing
open file descriptor for the device but want to wait for the release
fences to be signaled in the right time. The users can release their
open file descriptor of the device node created by m2m1shot2
immediately after M2M1SHOT2_PROCESS ioctl with non-block mode returns.

Holding mutex during m2m1shot2_cancel_context() is also removed
because the mutex is not locked at that time. m2m1shot2_context.mutex
is just designed to serialize ioctl calls.

Change-Id: I60f8a7500f77a70f40f066062ea68f11c501a047
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>