[8890] video: exynos: decon: Reconfig the window_update state if it is in partial mode.
If the WIN_CONFIG is failed, Reconfig the window_update state if it is in partial mode for
the next frame.
If window_update state is not reset, there will be a update register fail problem.
It is because of the mismatch between the win_update driver state and SFR state.

Change-Id: I2be99bcac12f964672fe7f4e9211b20e382d9098
Signed-off-by: SeWoon Park <seuni.park@samsung.com>
Signed-off-by: Meka Rama <v.meka@samsung.com>