[COMMON] cpufreq: interactive: fix sudden frequency drop even if has heavy task.
In the cluster-base dvfs, when some cores want to scale frequency
later than the other core in a policy, sometimes can't update target_freq for a long time
because above_hispeed_delay of a frequency already scaled by the other core in a policy.

core 0, 1 share a policy and in a cluster.
	 core 0				core 1
 task A is running
 change freq to 1.5GHz
				wake up from idle by task B
				target_freq is minimum freq(760MHz)
 change freq to 1.8GHz
				try change freq to 2.0GHz but can't scale
				because above_hispeed_delay by 1.8GHz
				so still target_freq is minimum freq(760MHz)
 finish task A. go to Idle
				try change freq to 2.0GHz but can't scale
				because above_hispeed_delay by 1.8GHz
				so still target_freq is minimum freq(760MHz)

 change freq to minimum(760)	still runnig task B, but sudden frequency
 becase wake up from idle.	drop happend by core 0 in the same policy.

To prevent this situation, updates target_freq to policy->cur
when governor operation is skipped by above hispeed delay.
this patch gives cores in a policy chance to update target freq.

Change-Id: Ic0ffdf4f71004411444e25a030bc712a146dc0da
Signed-off-by: Youngtae Lee <yt0729.lee@samsung.com>