proc: uid: Adds accounting for the cputimes per uid.
Adds proc files /proc/uid_cputime/show_uid_stat and
/proc/uid_cputime/remove_uid_range.

show_uid_stat lists the total utime and stime for the active as well as
terminated processes for each of the uids.

Writing a range of uids to remove_uid_range will delete the accounting
for all the uids within that range.

Change-Id: I5f16a41e4d60016f175080e6a28f13f8595d96e4
Signed-off-by: jinqian    <jinqian@google.com>
Signed-off-by: Jiseong Oh <jiseong.oh@samsung.com>