sched: cpufreq: Adds a field cpu_power in the task_struct
cpu_power has been added to keep track of amount of power each task is
consuming. cpu_power is updated whenever stime and utime are updated for
a task. power is computed by taking into account the frequency at which
the current core was running and the current for cpu actively
running at hat frequency.

Change-Id: I6765b38f3b9996432eac5c5c0005d45b9d0fa85e
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Signed-off-by: Jiseong Oh <jiseong.oh@samsung.com>