backlight: fix backlight control sequence
BUG=25867399
BUG=25635840
This fix the backlight issue: sometimes set to incorrect brightness
Brightness is set by the number of pulses input to bd82103.
If interrupt occured when making pulses the shape of pulse is corrupt
Then bd82103 detect incorrect number of pulses.
To avoid this I use spin_lock while making pulses.

Change-Id: Id7b25059a90b350f8fc985e165ec18c585934b63
Signed-off-by: Itsuki Yamashita <yamashitai@casio.co.jp>