[COMMON] media: mfc: check for available ctx on sleep
When mfc goes to sleep, the context is obtained from dev->curr_ctx.
There is a possibility that, last curr_ctx was closed and sleep was
issued before any activity happened on the remaining open context.
In that case, we get the below error on calling s5p_mfc_sleep :
s5p_mfc_sleep:526: no mfc context to run
This patch, will re-assign the dev->curr_ctx to a valid context so that
MFC sleep will not fail.

Change-Id: Ieb8be7bcd44ce08a73a0400455fe741dd5e12f1f
Signed-off-by: Jeonghee Kim <jhhhh.kim@samsung.com>