[8890] ASoC: samsung: change sram mapping method
We've mapped SRAM region when lpass driver is probed.
And we didn't unmap that region. But we mapped that region as
normal memory type. Speculative access is allowed at normal memory.
This means that it can potentially automatically load data into
the cache without the programmer having explicitly
requested a specific address.
So, we need to unmap sram region to prevent process from accessing
sram by speculative access.

Change-Id: I1a7d9feb4fc0be230c7ffcda88f755039613a19a
Signed-off-by: Hyunwoong Kim <khw0178.kim@samsung.com>