[COMMON] media: mfc: check and request UNPROT after DRC.
Some released frames can't be unprotected,
if there is no display buffer when the resolution is changed.
This buffers are unprotected in stop_streaming which is called after header parsing,
and it makes unprotection fail due to changed size.
So, Check and request unprotection in the end of DRC sequence.

Change-Id: I180c2faca1aa4a604d5a8d9d1bfff08a63941dbd
Signed-off-by: Jeonghee Kim <jhhhh.kim@samsung.com>