koi: audio: Modified the runtime pm sequence
For AK4678 codec, i2s lines are configured before the regulators are
enabled which gives rise to an inconsistent state for the LDO23 and I2C
SDA/SCL lines during resume of the drivers.

So the runtime sequence for the cpu_dai and codec_dai has been changed so
that codec resumes first.

Change-Id: I89c95f61d9ba8dcf45e4dbbca6699276fbdbbe2c
Signed-off-by: Chandni <ch.chandni@samsung.com>