input: Reduce unnecessary long wake_lock
After wake up by touch, it takes 20sec to suspend again.
Reduced to 3sec which is enough for not to disturb i2c communication
but shorter than screen timeout.
In addition, may fixed that in rare case AW goes into DOZE mode.
It seems that the palm-on event arrived before resuming completely
is cause of this issue, so masked this event 600ms after resuming.
note: wake_lock is necessary in the case of the following sequence:
- device suspend start
- touch irq occur
- device suspend complete

Change-Id: I1e303f0512d63cc0d71abfa0044d4be1a63ee64a
Signed-off-by: Shohei Sakamoto <sakamotos@casio.co.jp>