[COMMON] devfreq: governor timer is bound to cpu0
Timer of governor and thread
which is for devfreq timer are bound to cpu0.
Because if we do not bind timer and thread with particular
cpu which is included in boot cluster, timer and thread will
be running on cpu which is in nonboot cluster
although nonboot cluster is not running state.
It can cause increasing power consumption of cpu.
And add and delete comment at simple interactive governor.

Change-Id: Ia2076ffd61f7f0d83d1fb31017d3a45392ae097a
Signed-off-by: Jaejoon Yoo <joonyj7.yoo@samsung.com>