[7570] video: exynos: dsim: Error packet handling and stable read command
1. FIFO and interrupt are cleared before read request

2. The rx_err_handler goes into DSIM_CAL.

3. Using DSIM_CAL function when read RX_FIFO.

4. Some error bits is ignored in S.LSI AP.
DSI_PROTOCAL_VIOLATION[15] is for protocol violation that is caused EoTp
missing So this bit is egnored because of not supportung @S.LSI AP.
FALSE_ERROR_CONTROL[6] is for detect invalid escape or turnaround
sequence.
This bit is not supporting @S.LSI AP because of non standard
ULPS enter/exit sequence during power-gating.

Protocol violation that is caused EoTp missing
So this bit is egnored because of not supportung EoTp @ S.LSI AP

Change-Id: Ifc8a9d613ebcf14ddc1586316c38f3b4c7cfffae
Signed-off-by: Kim Manseok <manseoks.kim@samsung.com>