[COMMON] media: vb2-core: introduce verify_userptr memops
__qbuf_userptr() compares the address and the length given by the user
land with the previous successfully qbufed user memory area. But it
does not guarantee that the physical memory underneath the give user
address range is also the same because mmap after munmap may return
the same user address that is munmaped.
Therefore, vb2-core provides verify_userptr memops to give a chance
to the implementations of vb2 memops to check if the given user memory
range is really the same as the previously qbufed ones.

Change-Id: Idcd8c0b52a7c2a6f1578edb6f4a37bf7f1aa12c0
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>