RTC: delete the feature of time sync with subcpu
This feature is moved to system application.
Because rtc do not maintain milliseconds, it is hard to
syncronize the time in rtc driver with high accuracy.
The new application transfer the data of time to subcpu directly.
Please merge the application (newer than 4/15) in vendor/casio/koi
before check and merge this patch.

Change-Id: I97d7a882f3844fd32dbda69b40ee8f6b6150f5ce
Signed-off-by: Itsuki Yamashita <yamashitai@casio.co.jp>