[COMMON] set bus configuration to GSBUSCFG0 as cacheable/modifiable
From DWC USB3.0 DRD version 2.20A,
Bus configuration to select cache type was added.

In Exynos SoC, this cache type should be set as Cacheable/Modifiable.
But, this setting was not needed in SoCs before Exynos8890, because
AXI bus worked as Cacheable/Modifiable without the setting.

However in Exynos8890, it should be set as Cacheable/Modifiable.
If it is not set, USB will not work.

Change-Id: Ifcd514098a306a5dfefb6f5ccb8dcdbf3f15a40b
Signed-off-by: Lee Minho <minho55.lee@samsung.com>