[COMMON] usb: dwc3: code sync to adjust SoF accuracy
To increase the SOF accuracy of the USB3.0 DRD in host mode,
a feature to change the clock source used to generate SOF was added.
The clock source is changed from free clock to reference clock by this feature.
This change is a new feature of Link version 2.50A.

If this feature is not used, SOF accuracy meets the standrard.
However, some frameworks or applications require higher SoF accuracy for
synchronization with USB data traffic.
For these cases, this feature was added.

To use this feature, please add a member "adj-sof-accurac" to DT.

Change-Id: Ifb43261c68441486aff089ab730c85cbc5d04105
Signed-off-by: Lee Minho <minho55.lee@samsung.com>