[3250] iommy/exynos: remove sysmmu drivers for System MMU 6.x
The version of System MMUs of Exynos3250 varies between 1.x and 3.x.
The kernel 3.18 does not support the legacy System MMUs.
Instead of applying a log of patch series, this patch removes the
System MMU drivers entirely and copies the System MMU driver for
the 1.x and 3.x from the kernel 3.10.

Change-Id: I775c491c50a0c361e1bd9af0641d385c644698f6
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>