bcmdhd: Elevation of privilege vulnerability in Broadcom Wi-Fi driver
There is no validation of the batch_params.nchan variable in the
wls_parse_batching_cmd function leading to potential kernel memory corruption
when it is used in the dhd_dev_pno_set_for_batch function.
The fix is designed to add additional bounds checks to prevent the kernel
memory corruption.

Bug: 29009982
Change-Id: Ia23cce7fa6dbe9c6d3ac4cbc3c28c8ce42362aa0
Signed-off-by: Masakazu Yamazaki <yamazaki-m@casio.co.jp>