[COMMON] android: ion: deliver IOMMU prop to iommu driver properly
Fortunately, ion_iovmm_map() delivers the last argument to the iommu
driver, it is just an accident.
The parameters to ion_iovmm_map() should be delivered as the caller
has intended and it should be handled correctly. With the fix of the
argument passing, old legacy region_id handling is also removed.

Change-Id: I57d4463c56f1a73a0955bbe14b15ad940aa2dab5
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>