Fix denial of service vulnerability in kernel ext4 file system
Multiple race conditions in the ext4 filesystem implementation in the Linux kernel
before 4.5 allow local users to cause a denial of service (disk corruption) by
writing to a page that is associated with a different user's file after
unsynchronized hole punching and page-fault handling.
The fix is designed to disable support for the fallocate
FALLOC_FL_PUNCH_HOLE to prevent the race conditions.
A-28760453
CVE-2015-8839

Change-Id: I83afe9d006b3d3e31b182ecbbf5db32a15178e7c
Signed-off-by: Itsuki Yamashita <yamashitai@casio.co.jp>