[COMMON] trace: ess: prevent spinlock in panic
In panic sequence, we output all task information in this
kernel for debugging. In order to run it, all task should be
search with spinlock using task_lock API.
If lowmemkiller is run with searhing killed task, it used to
search all task with task_lock API. If Panic is occurred
at this time, lockup is possible to occurs.
This patch prevents it at that case.

Change-Id: I954195210a3a07fa4c440b557504bac48728f8c6
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>