[COMMON] arm64: arch_timer: enable ARM Errata 858921
This patch enables to apply a workaround about
ARM Errata 858921 as following.

CNTVCT and CNTPCT are updated according to the counter value
given by CNTVALUEB input bus. If CNtvct or CNTPCT are read while
they are updated with the lower 32 bits rolled-over, it is possible
to read a wrong transient value on CNTVCT or CNTPCT respoectively.

This can happen in AArch32 or in AArch64.

CNTVALUEB is incremented so that CNTVCT or CNTPCT have their 32 lower
bits rolled-over and their 32 higher bits incremented.
CNTVCT or CNTPCT are read when this update occurs.
There is one-cycle window when the read might return the old value for
the 32 lower bits of the counter and the new value for the 32 higher
bits.

Change-Id: I16b1bceb436c0cd13f4e74f32e7a4925c2f0bc08
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>